+++ source/Internal/Framework/Config/DataObject/ShopConfigurationSetting.php
@@ -11,7 +11,6 @@ namespace OxidEsales\EshopCommunity\Internal\Framework\Config\DataObject;
 
 class ShopConfigurationSetting
 {
-    public const MODULE_CLASS_EXTENSIONS_CHAIN = 'aModules';
     public const MODULE_CONTROLLERS = 'aModuleControllers';
     public const MODULE_SMARTY_PLUGIN_DIRECTORIES = 'moduleSmartyPluginDirectories';