+++ tests/Unit/Application/Model/SimpleVariantTest.php
@@ -24,7 +24,6 @@ class SimpleVariantTest extends \OxidTestCase
     protected function tearDown(): void
     {
         $this->cleanUpTable('oxdiscount');
-        oxTestModules::cleanAllModules();
         oxRegistry::get("oxDiscountList")->forceReload();
         $this->cleanUpTable('oxarticles');